Rename ovis_demo to ovis_driver
