+++ ovis_driver/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(ovis_demo)
+project(ovis_driver)
 
 add_compile_options(-std=c++11)
 find_package(catkin REQUIRED COMPONENTS
@@ -15,7 +15,7 @@ link_directories(${PROJECT_SOURCE_DIR}/lib)
 catkin_package( 
     INCLUDE_DIRS include
     CATKIN_DEPENDS kinova_driver kinova_msgs roscpp std_msgs geometry_msgs
-    LIBRARIES ovis_demo
+    LIBRARIES ovis_driver
 )
 
 include_directories(
@@ -23,7 +23,7 @@ include_directories(
   ${catkin_INCLUDE_DIRS}
 )
 
-add_executable(${PROJECT_NAME}_node src/ovis_demo.cpp include/ovis_demo.hpp include/kinova/Kinova.API.USBCommandLayerUbuntu.h)
+add_executable(${PROJECT_NAME}_node src/ovis_driver.cpp include/ovis_driver.hpp include/kinova/Kinova.API.USBCommandLayerUbuntu.h)
 
 add_dependencies(${PROJECT_NAME}_node ${catkin_EXPORTED_TARGETS})
