+++ ovis_driver/lib/USBCommLayerUbuntu.so
